# Conflicts:
#	backend/main-controller/pom.xml
#	backend/main-controller/src/main/java/com/adventureit/maincontroller/Controller/MainController.java
#	backend/microservices/adventure-service/src/main/java/com/adventureit/adventureservice/Requests/CreateAdventureRequest.java
#	backend/microservices/adventure-service/src/main/java/com/adventureit/adventureservice/Service/AdventureServiceImplementation.java

+++ backend/microservices/adventure-service/src/main/java/com/adventureit/adventureservice/Service/AdventureServiceImplementation.java
@@ -56,7 +56,7 @@ public class AdventureServiceImplementation implements AdventureService {
         else if (req.getName() == null){
             throw new NullFieldException("Create Adventure Request: Adventure Name NULL");
         }
-        Adventure persistedAdventure = this.adventureRepository.save(new Adventure(req.getName(),req.getDescription(), UUID.randomUUID(), req.getOwnerId(), req.getStartDate(), req.getEndDate()));
+        Adventure persistedAdventure = this.adventureRepository.save(new Adventure(req.getName(),req.getDescription(), UUID.randomUUID() , req.getOwnerId(), req.getStartDate(), req.getEndDate(),req.getLocation()));
         CreateAdventureResponse response = new CreateAdventureResponse(true);
         response.setAdventure(persistedAdventure);
         return response;
@@ -210,12 +210,21 @@ public class AdventureServiceImplementation implements AdventureService {
     }
 
     @Transactional
-    public RemoveAdventureResponse removeAdventure(UUID id){
+    public RemoveAdventureResponse removeAdventure(UUID id, UUID userID) throws Exception {
         Adventure retrievedAdventure = adventureRepository.findAdventureByAdventureId(id);
         if(retrievedAdventure == null){
             throw new AdventureNotFoundException("Remove Adventure: Adventure not found");
         }
-        adventureRepository.deleteAdventureByAdventureId(id);
+        if(!retrievedAdventure.getAttendees().contains(userID)){
+            throw new Exception("User does not belong to Adventure");
+        }
+
+        retrievedAdventure.getAttendees().remove(userID);
+
+        if(retrievedAdventure.getAttendees().isEmpty()){
+            adventureRepository.deleteAdventureByAdventureId(id);
+        }
+
         return new RemoveAdventureResponse(true, "Adventure successfully removed");
     }
 
@@ -232,11 +241,11 @@ public class AdventureServiceImplementation implements AdventureService {
     @Override
     public void mockPopulate(){
         final UUID mockOwnerID = UUID.fromString("1660bd85-1c13-42c0-955c-63b1eda4e90b");
-        final UUID mockAttendeeID = UUID.fromString("7a984756-16a5-422e-a377-89e1772dd71e");
+        final UUID mockAttendeeID = UUID.fromString("69e8eb21-eb63-4c83-9187-181a648bb759");
 
-        Adventure mockAdventure1 = new Adventure("Mock Adventure 1","Mock Description 1 Mock Description 1Mock Description 1 Mock Description 1 Mock Description 1 Mock Description", UUID.randomUUID(), mockOwnerID, LocalDate.of(2021, 7, 5),LocalDate.of(2021, 7, 9));
-        Adventure mockAdventure2 = new Adventure("Mock Adventure 2","Mock Description 2", UUID.randomUUID(), mockOwnerID, LocalDate.of(2021, 1, 3),LocalDate.of(2022, 1, 2));
-        Adventure mockAdventure3 = new Adventure("Mock Adventure 3", "Mock Description 3",UUID.randomUUID(), mockAttendeeID, LocalDate.of(2022, 1, 4),LocalDate.of(2022, 1, 22));
+        Adventure mockAdventure1 = new Adventure("Mock Adventure 1","Mock Description 1 Mock Description 1Mock Description 1 Mock Description 1 Mock Description 1 Mock Description", UUID.randomUUID(), mockOwnerID, LocalDate.of(2021, 7, 5),LocalDate.of(2021, 7, 9),UUID.randomUUID());
+        Adventure mockAdventure2 = new Adventure("Mock Adventure 2","Mock Description 2", UUID.randomUUID(), mockOwnerID, LocalDate.of(2021, 1, 3),LocalDate.of(2022, 1, 2),UUID.randomUUID());
+        Adventure mockAdventure3 = new Adventure("Mock Adventure 3", "Mock Description 3",UUID.randomUUID(), mockAttendeeID, LocalDate.of(2022, 1, 4),LocalDate.of(2022, 1, 22),UUID.randomUUID());
 
         mockAdventure1.addContainer(UUID.fromString("d53a7090-45f1-4eb2-953a-2258841949f8"));
         mockAdventure1.addContainer(UUID.fromString("26356837-f076-41ec-85fa-f578df7e3717"));
